add v2.1 into api(2.4)
